Rename folder and example/lib/main.dart to complex_example_main.dart
