+++ example/lib/complex_example_main.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+import 'complex_example_pages/main_page.dart';
+
+void main() => runApp(const MyApp());
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const MainPage();
+  }
+}
