+++ example/lib/complex_example_pages/palette_picker_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'package:flutter_hsvcolor_picker/flutter_hsvcolor_picker.dart';
 
 class PalettePickerPage extends StatefulWidget {
